requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
seaborn
matplotlib
tensorflow
keras
wordcloud
joblib
scikit-learn
keras